requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
httpx==0.28.1
yt-dlp==2026.2.20
youtube-transcript-api==1.2.3
